user-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
marshmallow==3.19.0
email-validator==2.0.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0